requirements.txt
numpy
streamlit
pymupdf
python-pptx